Changed checksumming algorithm
The checksum for saving and loading USFM data used to be based on the
number of words in the data. But for some reason Javascript did not
work well with this in case of more exotic UTF8 data. For that reason
the algorithm now uses the text length in bytes as the checksum.