Libgit2 configuration update
Bibledit used to compile and link against the system’s version of
libgit2, and it still does that, but it now gives preference to libgit
if it finds that library installed in ../../dependencies/libgit. This
enables compilation of bibledit for other operating systems and
architectures, for example iOS and Android.