lib: Super robust journal system
If there are too many files still left in the logbook folder, that
means that there's some problem that prevents the logbook entries from
being recorded in the database. This may lead to an infinite loop as
has been noticed at times, and this may quickly exhaust the available
inodes on the filesystem. Therefore if there are too many files still
left in the logbook folder, clear them out. This leads to loss of
logbook information, but be it so. Loss of information is better in
this case than a server that gets stuck. Also clear the logbook
database out.