lib: Passage navigator moved to menu bar.
The controls for the passage selector were too small to be operated by
finger on a touch-enabled device. The controls were fine for operation
by mouse click. The controls were moved to the menu bar, got the same
styles as the menu bar, and are now easily operated by touch.