lib: Can be started and stopped multiple times
The library now has functions for initialising it, starting and
stopping it, and shutting it down. Init and shutdown can be called once
during the lifetime of the app, and start and stop can be called
multiple times. This is for use on mobile apps, where an app can be
sent to the background, and then it should stop the lib, and when
moving to the foreground, it should start the library again.