lib: Javascript throws away topbar in iframe
Enable again the javascript bit that throws away the top bar. It used
to be there, was removed, but now appears to be direly needed again.
The topbar is removed by the server via the Workbench, but in other
cases the topbar makes it to the browser, and is to be removed here
within the iframe.