Summarizing sync messages
When sending/receiving every five minutes, the journal will be full of
standard messages saying it is sending and receiving, and then it is
ready. When the journal rotates, once an hour, it now summarizes all of
them into one general message. The standard journal entries are no
longer there after rotation, but the non-standard ones will remain
visible.