Fix notes database locked errors
There used to be many notes downloaders to run simultaneously. When
these downloaders happened to write to the database simultaneously,
there could be database locked errors in the journal. Now there is only
one notes downloader active at any time. It is expected that the locked
database errors go away.