Transfer raw passage from server to client
The reason for this is as follows: There is a difference between
Bibledit as written in PHP and Bibledit in C++, with regard to whether
the passages file has a new line at the end or not. To ensure that the
client passage file has exactly the same contents as on the server, the
contents should be passed in its raw form, without processing it. If it
were processed, then there will be a situation that the client keeps
downloading notes from the server, and never stops doing so, because
the passage file contents will always remain different. Raw passages
transfer fix this.