Fixed: Client deletes unsent note
Create a note on the client, while the client is sending and receiving:
Result is that the note on the client is deleted. The logbook shows
this. Thus the note is lost. The fix is to check for a note whether it
is unsent, before attempting to delete or modify it while downloading
notes from the server. It the note exists, it gets skipped.