Resources installation delays
Instead of interrupting an installation, and re-scheduling it, it would
be cleaner if the installation would delay a bit while the priority
tasks flag is on. That delay does not need to be visible in the
journal, but when it would be there, silently, then the effect is
obtained: To prioritize the other tasks - and at the same time the
journal is clean.