Allow 0 value to disable fading menu effect
The sliding menu that hides options after you (just barely) settle into
a screen drives me batty. I can see its usefulness in _some_ scenarios, but
not being able to disable it is crazy. It has annoyed all 3 people I've
sat down in front of the UI so far, so here's to giving users an
opt-out.