Only set flex-box on workspaces following top bar (fixes Safari)
Safari didn't like being in flex-box mode without an explicit height of
the overall workspace. It tried to squash all but the first box instead
of just using whatever space was available. This avoids setting flex-box
mode at all for pages that load in an iframe where the workspace is not
preceded by the topbar.