Move journal link to top of animated area
Fix #105

This is not really the correct fix, but it will tide people over so the
link isn't obscured by the journal content. The real problem is with the
animation code in Javascript, but the real fix is to scrap that hack and
to put this content (along with all other long scrolly pages) inside a
frame that scrolls just the content area.

As the latter is already addressed by the menu layout code changes I
have in the works, I think it's fair to patch this without redoing the
whole layout paradigm.