Work around pages with incomplete layout
See #128

Some pages (in the case of this bug, the notes content pages) use a
template that they don't actually have data for. If nothing gets put in
the top navigation bar, changes are the page is meant to scroll through
the whole page rather than be a flex layout.

To fix this properly these pages need to be converted so they either
don't use the same template or actually have a full layout specified.
Throwing data in containers without specifying what the layout should be
makes Safari go spastic.