Fix layout of non-workbench items in workspace
Fixes #111, #126, and #128

Some browsers we guessing the right thing, but Safari in particular
doesn't know how to deal with children of a flex box that don't have
properties. It assumes you want to jam them all in the same area, which
is what we want for the workbenches but any time we stuff other content
in the workspace we probably want each item to take up it's natural
space instead of getting squished.