Update scroller code for flex layouts
See #140

This change was made in 9c9bf20 but because the layout wasn't working in
Safari I tried a way to target some layouts and not others (bb84391). As
a result it had be be backed out (3bae7dd). That sort of worked but it
was ill-advised because it only worked for some pages in some browsers
and make things more complicated than they needed to be.

Commit c536402 fixed the actual flex box issue so all pages can use the
same element for scrollable content; 597eae1 removed the obsolete
workaround. Now we need to correctly target the right element to be
scrolled. Again.