Fix variable name mismatch introduced by b3ab9c54
See #141, fixes #139

This error is part of the reason scroll was jumping around and also the
cause of error message in Safari (through a chain reaction). The call to
a non existent variable only happened every other time the scroller was
called based on whether the element was or wasn't in view. With the old
scroller code being called twice for every event ("body,html"), it was
working "every time" because it was never run just once.