Drop negative margin from chapter number in default editor styles
Fixes #148

Negative margins for inline content are a bad idea. This was probably
added to correct for the following element being a paragraph and not
wanting browser default spacing, but the correct solution would have
been to remove it from the paragraph. Negative values here always break
something down the road. In this case they didn't break until we fixed
the other inter-paragraph spacing which obsoletes the need for this at
all.