Manual configuration of user-facing URL in a configuration file
It used to be possible to manually enter the user-facing URL through
the Bibledit interface. That worked fine as long as appropriate values
were entered. But after entering wrong values, it was not easy for a
user to navigate to the place where to correct the entered value.
Therefore the configuration will now be done through a text file on the
server. Incorrect values can now always be corrected easily. And it is
more secure.