Reads the public certificates and the private key from file
local.server.key: The private key for the server.
local.server.crt: The public certificate for the server.
local.authorities.crt: The chain of trusted certificate authorities
that have issued the server certificate. At the top of the file will be
the intermediate authority that issued the server certificate. Next can
be more intermediate authorities. At the bottom of the file should be
the trusted root certificate.