Removing redundant code for the Linux port
Building a tarball for Linux used to copy the Bibledit core library
sources into the Linux source tree. This led to redundant source in the
repository. After this change there’s no more redundant code in the
Linux source tree. The builder now builds the Linux tarball outside the
source tree, in a temporal location.