Trusted root certificates
The file is the same as the one that comes with cURL. The data is
downloaded from the Mozilla source tree, and then converted by a tool
included with cURL, into the cas.crt file, then the date is filtered
out, and it’s copied into the Bibledit source tree. This occurs every
night.