Fixed: Cannot scroll through resources in resource organizer
This fixes #187. Drag and drop, on a touch devices, those two do not
work together in the same way as they work together on a mouse operated
screen. For just now drag and drop has been disabled, and reordering
the resources work with buttons to move a resource up or down.