The visual verse editor puts all non-editable notes at the bottom of the chapter
This fixes #306.
User asks: Could all footnotes that don't belong to the focused verse
be placed at the bottom of the chapter?
The notes will be moved on the server in C++ with pugixml.
Unit tests for all logic done in the visual verse editor.
It tests new logic to be added to the page, in particular for moving
notes to the suffix, both in case there's already notes in the suffix,
and when there's no notes in the suffix, also when there's no notes in
the prefix.