The settings for the permitted change in the Bible editors adapt to the enabled editors
This fixes #278

Phil: [Having fewer Bible editors.] This would mean that the Settings >
Personalize change percent settings might be changed to refer to just
Visual vs. USFM editors. This would make the settings easier to
understand.

Teus: When the Bible editors are going to be configurable, it is good
to adapt the setting to that, so that the whole is consistent to the
user.

What is to be done is to use the names as they are in the menu,
exactly, or use two names.

If there's no editors for a category, the setting won't be visible.