Keyboard shortcut Ctrl-\ to cycle through the available Bible editors
This fixes #279:

Phil: Finally, if there were only two editors for the desktops, it
would be great to have a hot key to switch between them.

Teus: Even with more than two editors, or with only two configured to
be active, in all cases the hot key can be made, and it would cycle
through the available configured editors.

The same as in Bibledit-Gtk: Ctrl-\

It calls the URL on the server, for switching to the "next" editor, and
that URL on the server determines which one the next will be, then
forwards the page to it.