Fixing database errors
SQlite is supposed to be multi-threaded if compiled as such. But in
rare cases it still gave database errors. Therefore the Bibledit
library is now providing its own mutexes on top of whatever SQLite
already has. This proves to be the solution, because the database locks
and errors have not been noticed again, yet ). This fixes #284