IPv6 #292
There’s now two http servers, one listening for IPv4 clients, and
another one listening for IPv6 clients. There is a way to have one
server listen for both IPv4 and IPv6, but then it rewrites the IPv4
address as an IPv6 address, and that would be confusing, so it is
clearer and better to have two servers.