One plain http server rather than two #292
It used to have two servers one listening on IPv4 and the other on
IPv6. That worked fine on BSD. But on Linux, the IPv6 server failed to
bind to the socket once the IPv4 server was up. The solution was to
have one server listening for IPv6 and IPv4.