Safer limits on the cache size in the Cloud
Limit cache size #363

There have been instances that the cache takes up 4, 5, 6 Gbytes in the
Cloud. And this can be more than that even. This may lead to errors
when the disk runs out of space. Therefore it's good to limit the cache
size in the Cloud. Currently the cache is cleared after a certain time,
but this change implements also constraining the cache size based on
size.