Fixed: The app icon set references a file but that file does not have a valid extension #385
In Xcode 7++ building the project gives this warning:

"The app icon set ... references a file ... but that file does not have
a valid extension"

It appears that changing the file to .png or even leaving out the
extension at all, fixes the warning.

The idea is to fix all warnings of this type.