Implemented: Paragraph capital check needs more context sensitivity
#393
The check for capitals at the start of paragraphs is quite useful, but
it is throwing a ton of false positives for us because we are not
capitalizing the first character in lines of poetry that are
continuations of previews sentences. Each sentence gets a capital, but
not each line of poetry. I think there could be a user configurable
toggle for whether or not \q1 and \q2 tags count as "starting a
paragraph" for the purposes of this check.