New files for catholic versification
This fixes #401:
I discovered that the Catholic 1 into the versification propositions,
was completely false. For this reason, I took as model the Catholic 2
from Sword. I made a new versication for bibledit web Catholic2.txt.
I send you every thing, that you can add it to the software.
One new versification (Catholic1, as the Catholic on Sword, that means
Esther with 10 chapters) and other files with new corrections.